refactor: Adding in import-resolver-node to help with the import paths and eslint.
Updating the state of the books to be the same on the bookshelf as well as the search page.

Moving the main API call to fetch all of the books for current user to the master App component. Storing the data in state and sprinkling it down to the child components.

Adding the refetchAllBooks call to the un-mounting of the BookSearch component.